fastapi/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-dotenv>=1.0.0
openai>=1.0.0
moviepy>=1.0.0
pillow>=9.0.0
requests